move layout folder into components
