+++ frontend/package.json
@@ -38,7 +38,7 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "dev": "next dev --webpack",
+    "dev": "next dev",
     "start": "next dev",
     "build": "next build",
     "lint": "next lint",
