+++ frontend/src/App.css
@@ -1,3 +1,16 @@
+.layout-flex-wrapper {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
+.govuk-main-wrapper {
+  flex: 1 0 auto;
+}
+
+.govuk-footer {
+  flex-shrink: 0;
+}
 .App {
   text-align: center;
 }
