+++ frontend/src/components/InvalidFieldsTable.tsx
