+++ frontend/src/components/layout/CookieBanner.tsx
