+++ frontend/src/components/layout/HelpAndSupport.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import {TITLE} from '../utils/Constants';
+import {serviceName} from '../../utils/Constants';
 
 const HelpAndSupport: React.FC = () =>{
 
@@ -8,7 +8,7 @@ const HelpAndSupport: React.FC = () =>{
         <div>
         <p className="govuk-heading-m">Help and support</p>
         <p className="govuk-body">    
-        If you are having problems, please contact the {TITLE} service team via this link: <Link to="/contact-us">Contact us</Link>
+        If you are having problems, please contact the {serviceName} service team via this link: <Link to="/contact-us">Contact us</Link>
         </p>
         </div>
