+++ frontend/src/components/layout/NotificationBanner.tsx
