+++ frontend/src/components/layout/PhaseBanner.tsx
