+++ frontend/src/views/contactUs.tsx
@@ -1,8 +1,7 @@
 import React, {  useContext} from "react";
-import { TwoThirdsLayout } from "../Layout/Layout";
+import { TwoThirdsLayout } from "../components/layout/Layout";
 import {IsLoggedInContext} from "../utils/login/LoginProvider";
-import Footer from "../Layout/Footer";
-import { TITLE } from "../utils/Constants";
+import { serviceName } from "../utils/Constants";
 
 
 const ContactUs: React.FC = () => {
@@ -15,7 +14,7 @@ const ContactUs: React.FC = () => {
     <>
       <TwoThirdsLayout title="Contact Us" description="Home" hideCookieBanner={true} isLoggedIn={isLoggedIn}>
         <h1 className="govuk-heading-xl">
-        Contact the {TITLE.split(" ")
+        Contact the {serviceName.split(" ")
                           .slice(1)
                           .join(" ")} service team
         </h1>
@@ -24,7 +23,7 @@ const ContactUs: React.FC = () => {
         </p>
         <div className="govuk-!-font-size-19 govuk-!-margin-bottom-9">
         <p className="govuk-body">
-        If you are experiencing technical issues, or if you have any suggestions, comments or criticisms, please contact the {TITLE.split(" ").slice(1).join(" ")} service team through one of the channels below.
+        If you are experiencing technical issues, or if you have any suggestions, comments or criticisms, please contact the {serviceName.split(" ").slice(1).join(" ")} service team through one of the channels below.
         </p>
         <p className="govuk-body">
         The Help Desk is available Monday to Friday, 9am to 5pm <br />(excluding Bank Holidays in England and Wales, and the 24th of December).
@@ -38,7 +37,6 @@ const ContactUs: React.FC = () => {
         </p>
         </div>
       </TwoThirdsLayout>
-      <Footer />
     </>
   );
 };
