+++ frontend/src/views/cookiePage.tsx
@@ -1,8 +1,7 @@
 import React, { useContext } from "react";
-import { TwoThirdsLayout } from "../Layout/Layout";
+import { TwoThirdsLayout } from "../components/layout/Layout";
 import { IsLoggedInContext } from "../utils/login/LoginProvider";
-import Footer from "../Layout/Footer";
-import { TITLE } from "../utils/Constants";
+import { serviceName } from "../utils/Constants";
 
 const CookiePage: React.FC = () => {
   const { isLoggedIn } = useContext(IsLoggedInContext);
@@ -16,7 +15,7 @@ const CookiePage: React.FC = () => {
         isLoggedIn={isLoggedIn}
       >
         <h1 className="govuk-heading-xl">
-          Cookies on the {TITLE} Service
+          Cookies on the {serviceName} Service
         </h1>
         <div className="govuk-!-font-size-19">
           <p className="govuk-body">
@@ -70,7 +69,6 @@ const CookiePage: React.FC = () => {
           </table>
         </div>
       </TwoThirdsLayout>
-      <Footer />
     </>
   );
 };
