+++ frontend/src/views/registrations.tsx
@@ -1,10 +1,9 @@
 import React, { useContext } from "react";
-import { FullColumnLayout } from "../Layout/Layout";
+import { FullColumnLayout } from "../components/layout/Layout";
 import { Link } from "react-router-dom";
 import { IsLoggedInContext } from "../utils/login/LoginProvider";
-import Footer from "../Layout/Footer";
 import { v4 as uuidv4 } from "uuid";
-import { TITLE } from "../utils/Constants";
+import { serviceName } from "../utils/Constants";
 const Registration: React.FC = () => {
   const { isLoggedIn } = useContext(IsLoggedInContext);
     const LINKS = [
@@ -30,7 +29,7 @@ const Registration: React.FC = () => {
         hideCookieBanner={true}
         isLoggedIn={isLoggedIn}
       >
-        <h1 className="govuk-heading-xl">{TITLE}</h1>
+        <h1 className="govuk-heading-xl">{serviceName}</h1>
         {LINKS.map((value) => {
           return (
             <Link
@@ -43,7 +42,6 @@ const Registration: React.FC = () => {
           );
         })}
       </FullColumnLayout>
-      <Footer />
     </>
   );
 };
